docs: change in README.md structure to add information about the project
Changes were made in the document to add the necessary information
about the project in the following way, adding the title and elaboration
of the project, a brief description about the project, its development
and user's use. The data used for its elaboration and the necessary
aids for the realization of the source code used.